First upload
The AI currently appears to perhaps be able to improve however because of the type of back propagation used it would take a long time when it is not only dealing with a single letter. The letters files are needed to run the AI or others can be created using a different font etc. It is also possible to create an AI that uses are far bigger grid or can recognise  more than 26 letters.